pyproject.toml
[tool.poetry]
name = "skit-labels"
version = "0.3.1"
description = "Command line tool for interacting with labelled datasets at skit.ai."
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^19.3.0"
pydash = "^5.0.2"
docopt = "^0.6.2"
tqdm = "^4.46.0"
psycopg2-binary = "^2.8.5"
pytz = "^2020.1"
python-dateutil = "^2.8.1"
toml = "^0.10.2"
skit-fixdf = "^0.1.0"
dvc = {extras = ["s3"], version = "^2.9.3"}
loguru = "^0.5.3"
jsonschema = "^4.4.0"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"

[tool.poetry.scripts]
skit-labels = "skit_labels.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
